Update testcase of parser
