+++ src/test/java/seedu/address/logic/parser/AddModuleToStudentTakenCommandParserTest.java
@@ -1,10 +1,11 @@
 package seedu.address.logic.parser;
 
-import org.junit.Test;
-import seedu.address.logic.commands.AddModuleToStudentTakenCommand;
-
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
+
+import org.junit.Test;
+
+import seedu.address.logic.commands.AddModuleToStudentTakenCommand;
 
 public class AddModuleToStudentTakenCommandParserTest {
     private AddModuleToStudentTakenCommandParser parser = new AddModuleToStudentTakenCommandParser();
