+++ src/test/java/seedu/address/logic/parser/RemoveModuleFromStudentStagedCommandParserTest.java
@@ -1,18 +1,20 @@
 package seedu.address.logic.parser;
 
-import org.junit.Test;
-import seedu.address.logic.commands.AddModuleToStudentStagedCommand;
-
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 
-public class AddModuleToStudentStagedCommandParserTest {
-    private AddModuleToStudentStagedCommandParser parser = new AddModuleToStudentStagedCommandParser();
+import org.junit.Test;
+
+import seedu.address.logic.commands.RemoveModuleFromStudentStagedCommand;
+
+
+public class RemoveModuleFromStudentStagedCommandParserTest {
+    private RemoveModuleFromStudentStagedCommandParser parser = new RemoveModuleFromStudentStagedCommandParser();
 
     @Test
     public void parse_emptyCode_failure() {
         String expectedMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                AddModuleToStudentStagedCommand.MESSAGE_USAGE);
+                RemoveModuleFromStudentStagedCommand.MESSAGE_USAGE);
         assertParseFailure(parser, "", expectedMessage);
     }
