+++ src/test/java/seedu/address/logic/parser/RemoveModuleFromStudentTakenCommandParserTest.java
@@ -1,19 +1,19 @@
 package seedu.address.logic.parser;
 
-import org.junit.Test;
-import seedu.address.logic.commands.AddModuleToStudentStagedCommand;
-import seedu.address.logic.commands.RemoveModuleFromStudentStagedCommand;
-
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 
-public class RemoveModuleFromStudentStagedCommandParserTest {
-    private RemoveModuleFromStudentStagedCommandParser parser = new RemoveModuleFromStudentStagedCommandParser();
+import org.junit.Test;
+
+import seedu.address.logic.commands.RemoveModuleFromStudentTakenCommand;
+
+public class RemoveModuleFromStudentTakenCommandParserTest {
+    private RemoveModuleFromStudentTakenCommandParser parser = new RemoveModuleFromStudentTakenCommandParser();
 
     @Test
     public void parse_emptyCode_failure() {
         String expectedMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                RemoveModuleFromStudentStagedCommand.MESSAGE_USAGE);
+                RemoveModuleFromStudentTakenCommand.MESSAGE_USAGE);
         assertParseFailure(parser, "", expectedMessage);
     }
