+++ src/test/java/seedu/address/logic/parser/SearchCommandParserTest.java
@@ -1,4 +1,3 @@
-
 package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
@@ -9,27 +8,27 @@ import java.util.Arrays;
 
 import org.junit.Test;
 
-import seedu.address.logic.commands.FindCommand;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.logic.commands.SearchCommand;
+import seedu.address.model.module.CodeStartsKeywordsPredicate;
 
-public class FindCommandParserTest {
+public class SearchCommandParserTest {
 
-    private FindCommandParser parser = new FindCommandParser();
+    private SearchCommandParser parser = new SearchCommandParser();
 
     @Test
     public void parse_emptyArg_throwsParseException() {
-        assertParseFailure(parser, "     ", String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
+        assertParseFailure(parser, "     ", String.format(MESSAGE_INVALID_COMMAND_FORMAT, SearchCommand.MESSAGE_USAGE));
     }
 
     @Test
     public void parse_validArgs_returnsFindCommand() {
         // no leading and trailing whitespaces
-        FindCommand expectedFindCommand =
-                new FindCommand(new NameContainsKeywordsPredicate(Arrays.asList("Alice", "Bob")));
-        assertParseSuccess(parser, "Alice Bob", expectedFindCommand);
+        SearchCommand expectedSearchCommand =
+                new SearchCommand(new CodeStartsKeywordsPredicate(Arrays.asList("CS", "ACC")));
+        assertParseSuccess(parser, "CS ACC", expectedSearchCommand);
 
         // multiple whitespaces between keywords
-        assertParseSuccess(parser, " \n Alice \n \t Bob  \t", expectedFindCommand);
+        assertParseSuccess(parser, " \n CS \n \t ACC  \t", expectedSearchCommand);
     }
 
 }
